requirements.txt
tensorflow==2.19.0
pandas 
dvc
mlflow==2.2.2
notebook
numpy
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
scipy
Flask
Flask-Cors
gdown
streamlit 
streamlit-lottie 
pillow
fpdf
#-e .